Remove WinRT inheritance from PaletteItem
Right now, we construct **two objects** for every palette item: the
derived type and the base type. It's unnnecessary.

This also removes many of our palette items from our IDL. The only ones
that are necessary to expose via our WinRT API surface are the ones that
are used in XAML documents.