Enable DECRQCRA at compile time for all Terminal builds
Support for DECRQCRA Request Checksum of Rectangular Area was added
in #14989, but left disabled at build time because it could be
considered a security risk.

In #17895, we unconditionally added it to Terminal's settings UI. For
users on Stable and Preview, it didn't actually enable anything. Whoops.

Since we have a way to turn it off (and in so doing, mitigate the risk)
in Terminal, it's high time for us to remove the feature gating.

Conhost doesn't support turning it off for now and so conhost can still
have it compiled out, as a treat.